# auth/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auth",
    "summary": "Сервис авторизации онлайн-кинотеатра",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/api/v1/roles": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Список ролей",
        "operationId": "list_roles_auth_api_v1_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleSchema"
                  },
                  "type": "array",
                  "title": "Response List Roles Auth Api V1 Roles Get"
                }
              }
            }
          },
          "503": {
            "description": "PostgreSQL или Redis временно недоступны",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "example": {
                  "code": "service_unavailable",
                  "detail": "Service temporarily unavailable, retry later"
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`, `token_expired`, `token_invalid`, `token_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "examples": {
                  "not_authenticated": {
                    "summary": "Authorization header with a bearer access token is required",
                    "value": {
                      "code": "not_authenticated",
                      "detail": "Authorization header with a bearer access token is required"
                    }
                  },
                  "token_expired": {
                    "summary": "Token has expired",
                    "value": {
                      "code": "token_expired",
                      "detail": "Token has expired"
                    }
                  },
                  "token_invalid": {
                    "summary": "Token is malformed or its signature is invalid",
                    "value": {
                      "code": "token_invalid",
                      "detail": "Token is malformed or its signature is invalid"
                    }
                  },
                  "token_revoked": {
                    "summary": "Session has been terminated, log in again",
                    "value": {
                      "code": "token_revoked",
                      "detail": "Session has been terminated, log in again"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "`permission_denied`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "examples": {
                  "permission_denied": {
                    "summary": "Not enough permissions",
                    "value": {
                      "code": "permission_denied",
                      "detail": "Not enough permissions"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "roles"
        ],
        "summary": "Создание роли",
        "operationId": "create_role_auth_api_v1_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSchema"
                }
              }
            }
          },
          "503": {
            "description": "PostgreSQL или Redis временно недоступны",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "example": {
                  "code": "service_unavailable",
                  "detail": "Service temporarily unavailable, retry later"
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`, `token_expired`, `token_invalid`, `token_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "examples": {
                  "not_authenticated": {
                    "summary": "Authorization header with a bearer access token is required",
                    "value": {
                      "code": "not_authenticated",
                      "detail": "Authorization header with a bearer access token is required"
                    }
                  },
                  "token_expired": {
                    "summary": "Token has expired",
                    "value": {
                      "code": "token_expired",
                      "detail": "Token has expired"
                    }
                  },
                  "token_invalid": {
                    "summary": "Token is malformed or its signature is invalid",
                    "value": {
                      "code": "token_invalid",
                      "detail": "Token is malformed or its signature is invalid"
                    }
                  },
                  "token_revoked": {
                    "summary": "Session has been terminated, log in again",
                    "value": {
                      "code": "token_revoked",
                      "detail": "Session has been terminated, log in again"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "`permission_denied`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "examples": {
                  "permission_denied": {
                    "summary": "Not enough permissions",
                    "value": {
                      "code": "permission_denied",
                      "detail": "Not enough permissions"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "`role_name_taken`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                },
                "examples": {
                  "role_name_taken": {
                    "summary": "Role with this name already exists",
                    "value": {
                      "code": "role_name_taken",
                      "detail": "Role with this name already exists"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ]
      }
    },
    "/auth/api/v1/roles/{role_id}": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Роль",
        "operationId": "get_role_auth_api_v1_roles__role_id__get",
        "security": [
          {
            "accessToken": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSchema"
                }
              }
            }
          },
          "503": {
            "description": "PostgreSQL или Redis временно недоступны",
            "content": {
              "application/json": {
                "example": {
                  "code": "service_unavailable",
                  "detail": "Service temporarily unavailable, retry later"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`, `token_expired`, `token_invalid`, `token_revoked`",
            "content": {
              "application/json": {
                "examples": {
                  "not_authenticated": {
                    "summary": "Authorization header with a bearer access token is required",
                    "value": {
                      "code": "not_authenticated",
                      "detail": "Authorization header with a bearer access token is required"
                    }
                  },
                  "token_expired": {
                    "summary": "Token has expired",
                    "value": {
                      "code": "token_expired",
                      "detail": "Token has expired"
                    }
                  },
                  "token_invalid": {
                    "summary": "Token is malformed or its signature is invalid",
                    "value": {
                      "code": "token_invalid",
                      "detail": "Token is malformed or its signature is invalid"
                    }
                  },
                  "token_revoked": {
                    "summary": "Session has been terminated, log in again",
                    "value": {
                      "code": "token_revoked",
                      "detail": "Session has been terminated, log in again"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "403": {
            "description": "`permission_denied`",
            "content": {
              "application/json": {
                "examples": {
                  "permission_denied": {
                    "summary": "Not enough permissions",
                    "value": {
                      "code": "permission_denied",
                      "detail": "Not enough permissions"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "404": {
            "description": "`role_not_found`",
            "content": {
              "application/json": {
                "examples": {
                  "role_not_found": {
                    "summary": "Role not found",
                    "value": {
                      "code": "role_not_found",
                      "detail": "Role not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "roles"
        ],
        "summary": "Изменение роли",
        "description": "Меняет переданные поля. Новые права действуют сразу у всех пользователей с этой ролью.",
        "operationId": "update_role_auth_api_v1_roles__role_id__patch",
        "security": [
          {
            "accessToken": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSchema"
                }
              }
            }
          },
          "503": {
            "description": "PostgreSQL или Redis временно недоступны",
            "content": {
              "application/json": {
                "example": {
                  "code": "service_unavailable",
                  "detail": "Service temporarily unavailable, retry later"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`, `token_expired`, `token_invalid`, `token_revoked`",
            "content": {
              "application/json": {
                "examples": {
                  "not_authenticated": {
                    "summary": "Authorization header with a bearer access token is required",
                    "value": {
                      "code": "not_authenticated",
                      "detail": "Authorization header with a bearer access token is required"
                    }
                  },
                  "token_expired": {
                    "summary": "Token has expired",
                    "value": {
                      "code": "token_expired",
                      "detail": "Token has expired"
                    }
                  },
                  "token_invalid": {
                    "summary": "Token is malformed or its signature is invalid",
                    "value": {
                      "code": "token_invalid",
                      "detail": "Token is malformed or its signature is invalid"
                    }
                  },
                  "token_revoked": {
                    "summary": "Session has been terminated, log in again",
                    "value": {
                      "code": "token_revoked",
                      "detail": "Session has been terminated, log in again"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "403": {
            "description": "`permission_denied`",
            "content": {
              "application/json": {
                "examples": {
                  "permission_denied": {
                    "summary": "Not enough permissions",
                    "value": {
                      "code": "permission_denied",
                      "detail": "Not enough permissions"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "404": {
            "description": "`role_not_found`",
            "content": {
              "application/json": {
                "examples": {
                  "role_not_found": {
                    "summary": "Role not found",
                    "value": {
                      "code": "role_not_found",
                      "detail": "Role not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "409": {
            "description": "`role_name_taken`",
            "content": {
              "application/json": {
                "examples": {
                  "role_name_taken": {
                    "summary": "Role with this name already exists",
                    "value": {
                      "code": "role_name_taken",
                      "detail": "Role with this name already exists"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "roles"
        ],
        "summary": "Удаление роли",
        "description": "Удаляет роль и отбирает её у всех пользователей.",
        "operationId": "delete_role_auth_api_v1_roles__role_id__delete",
        "security": [
          {
            "accessToken": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "503": {
            "description": "PostgreSQL или Redis временно недоступны",
            "content": {
              "application/json": {
                "example": {
                  "code": "service_unavailable",
                  "detail": "Service temporarily unavailable, retry later"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`, `token_expired`, `token_invalid`, `token_revoked`",
            "content": {
              "application/json": {
                "examples": {
                  "not_authenticated": {
                    "summary": "Authorization header with a bearer access token is required",
                    "value": {
                      "code": "not_authenticated",
                      "detail": "Authorization header with a bearer access token is required"
                    }
                  },
                  "token_expired": {
                    "summary": "Token has expired",
                    "value": {
                      "code": "token_expired",
                      "detail": "Token has expired"
                    }
                  },
                  "token_invalid": {
                    "summary": "Token is malformed or its signature is invalid",
                    "value": {
                      "code": "token_invalid",
                      "detail": "Token is malformed or its signature is invalid"
                    }
                  },
                  "token_revoked": {
                    "summary": "Session has been terminated, log in again",
                    "value": {
                      "code": "token_revoked",
                      "detail": "Session has been terminated, log in again"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "403": {
            "description": "`permission_denied`",
            "content": {
              "application/json": {
                "examples": {
                  "permission_denied": {
                    "summary": "Not enough permissions",
                    "value": {
                      "code": "permission_denied",
                      "detail": "Not enough permissions"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "404": {
            "description": "`role_not_found`",
            "content": {
              "application/json": {
                "examples": {
                  "role_not_found": {
                    "summary": "Role not found",
                    "value": {
                      "code": "role_not_found",
                      "detail": "Role not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/api/v1/users/{user_id}/roles": {
      "get": {
        "tags": [
          "access"
        ],
        "summary": "Роли пользователя",
        "operationId": "user_roles_auth_api_v1_users__user_id__roles_get",
        "security": [
          {
            "accessToken": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleSchema"
                  },
                  "title": "Response User Roles Auth Api V1 Users  User Id  Roles Get"
                }
              }
            }
          },
          "503": {
            "description": "PostgreSQL или Redis временно недоступны",
            "content": {
              "application/json": {
                "example": {
                  "code": "service_unavailable",
                  "detail": "Service temporarily unavailable, retry later"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`, `token_expired`, `token_invalid`, `token_revoked`",
            "content": {
              "application/json": {
                "examples": {
                  "not_authenticated": {
                    "summary": "Authorization header with a bearer access token is required",
                    "value": {
                      "code": "not_authenticated",
                      "detail": "Authorization header with a bearer access token is required"
                    }
                  },
                  "token_expired": {
                    "summary": "Token has expired",
                    "value": {
                      "code": "token_expired",
                      "detail": "Token has expired"
                    }
                  },
                  "token_invalid": {
                    "summary": "Token is malformed or its signature is invalid",
                    "value": {
                      "code": "token_invalid",
                      "detail": "Token is malformed or its signature is invalid"
                    }
                  },
                  "token_revoked": {
                    "summary": "Session has been terminated, log in again",
                    "value": {
                      "code": "token_revoked",
                      "detail": "Session has been terminated, log in again"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "403": {
            "description": "`permission_denied`",
            "content": {
              "application/json": {
                "examples": {
                  "permission_denied": {
                    "summary": "Not enough permissions",
                    "value": {
                      "code": "permission_denied",
                      "detail": "Not enough permissions"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "404": {
            "description": "`user_not_found`",
            "content": {
              "application/json": {
                "examples": {
                  "user_not_found": {
                    "summary": "User not found",
                    "value": {
                      "code": "user_not_found",
                      "detail": "User not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorSchema": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Код ошибки для программ",
            "examples": [
              "token_expired"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Описание для человека",
            "examples": [
              "Access token has expired"
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorSchema",
        "description": "Ошибка."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleCreateSchema": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z][a-z0-9_-]*$",
            "title": "Name",
            "description": "Уникальное имя: латиница в нижнем регистре, цифры, _ и -",
            "examples": [
              "subscribers"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Описание"
          },
          "permissions": {
            "items": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[a-z][a-z0-9_]*(\\.[a-z][a-z0-9_]*)+$",
              "examples": [
                "films.subscription"
              ]
            },
            "type": "array",
            "maxItems": 100,
            "title": "Permissions",
            "description": "Права роли",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RoleCreateSchema",
        "description": "Новая роль."
      },
      "RoleSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Идентификатор"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Имя роли",
            "examples": [
              "subscribers"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Описание",
            "examples": [
              "Подписчики: фильмы по подписке"
            ]
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions",
            "description": "Права роли",
            "examples": [
              [
                "films.subscription"
              ]
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Когда создана"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Когда изменена"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "permissions",
          "created_at",
          "updated_at"
        ],
        "title": "RoleSchema",
        "description": "Роль и её права."
      },
...
      "RoleUpdateSchema": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 2,
                "pattern": "^[a-z][a-z0-9_-]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Новое имя",
            "examples": [
              "subscribers"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Новое описание"
          },
          "permissions": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 128,
                  "pattern": "^[a-z][a-z0-9_]*(\\.[a-z][a-z0-9_]*)+$",
                  "examples": [
                    "films.subscription"
                  ]
                },
                "type": "array",
                "maxItems": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Permissions",
            "description": "Новый список прав целиком"
          }
        },
        "type": "object",
        "title": "RoleUpdateSchema",
        "description": "Изменение роли: меняются только переданные поля."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "accessToken": {
        "type": "http",
        "description": "access-токен из ответа на вход или обновление токенов",
        "scheme": "bearer"
      }
    }
  },
...
}